+++ lib/cmdb.rb
@@ -134,9 +134,10 @@ module CMDB
   end
 end
 
-require 'cmdb/consul_source'
 require 'cmdb/environment_source'
 require 'cmdb/file_source'
+require 'cmdb/network_source'
+require 'cmdb/consul_source'
 require 'cmdb/interface'
 require 'cmdb/rewriter'
 require 'cmdb/commands'