+++ lib/cmdb/commands/shim.rb
@@ -102,13 +102,6 @@ Where [options] are selected from:
       @signal          = options[:reload_signal]
       @env             = options[:env]
       @user            = options[:user]
-      @root            = options[:root]
-
-      CMDB::FileSource.base_directories = @keys unless @keys.empty?
-      CMDB::ConsulSource.url = @consul_url unless @consul_url.nil?
-      if !@consul_prefixes.nil? && !@consul_prefixes.empty?
-        CMDB::ConsulSource.prefixes = @consul_prefixes
-      end
 
       CMDB.log.level = Logger::FATAL if options[:quiet]
     end
@@ -117,7 +110,7 @@ Where [options] are selected from:
     #
     # @raise [SystemExit] if something goes wrong
     def run
-      @cmdb = CMDB::Interface.new(root: @root)
+      @cmdb = CMDB::Interface.new
 
       rewrote   = rewrite_files
       populated = populate_environment