+++ lib/cmdb/consul_source.rb
@@ -1,83 +1,62 @@
 # encoding: utf-8
-require 'diplomat'
+require 'base64'
+require 'net/http'
+require 'open-uri'
 
 module CMDB
-  class ConsulSource
+  class ConsulSource < NetworkSource
     # Regular expression to match array values
     ARRAY_VALUE = /^\[(.*)\]$/
 
-    ### Class variables
-
-    class << self
-      attr_writer :url
-    end
-
-    class << self
-      attr_reader :url
-    end
-
-    class << self
-      attr_writer :prefixes
-    end
-
-    class << self
-      attr_reader :prefixes
-    end
-
-    ### Instance variables
-
-    # The url to communicate with consul
-    @url = nil
-
-    # Initialize the configuration for consul source
-    def initialize(prefix)
-      Diplomat.configure do |config|
-        config.url = self.class.url
-      end
-      @prefix = prefix
-    end
-
     # Get a single key from consul
     def get(key)
-      value = Diplomat::Kv.get(dot_to_slash(key))
-      process_value(value)
-    rescue TypeError
-      puts 'hi'
-    rescue Diplomat::KeyNotFound
-      nil
+      key = dot_to_slash(key)
+      response = http_get path_to(key)
+      case response
+      when Array
+        item = response.first
+        process_value(Base64.decode64(item['Value']))
+      when 404
+        nil
+      else
+        raise CMDB:Error.new("Unexpected consul response #{value.inspect}")
+      end
     end
 
-    # Iterate through all keys with a given prefix in consul
+    # Set a single key in consul.
+    def set(key, value)
+      key = dot_to_slash(key)
+      http_put path_to(key), value
+    end
+
+    # Iterate through all keys in this source.
+    # @return [Integer] number of key/value pairs that were yielded
     def each_pair(&_block)
-      prefix = @prefix || ''
-      all = Diplomat::Kv.get(prefix, recurse: true)
+      all = http_get path_to('/'), query:'recurse'
+      unless all.is_a?(Array)
+        raise CMDB:Error.new("Unexpected consul response #{all.inspect}")
+      end
+
       all.each do |item|
-        dotted_prefix = prefix.split('/').join('.')
-        dotted_key = item[:key].split('/').join('.')
+        dotted_prefix = @prefix.split('/').join('.')
+        dotted_key = item['Key'].split('/').join('.')
         key = dotted_prefix == '' ? dotted_key : dotted_key.split("#{dotted_prefix}.").last
-        value = process_value(item[:value])
+        value = process_value(Base64.decode64(item['Value']))
         yield(key, value)
       end
-    rescue Diplomat::KeyNotFound => exc
-      CMDB.log.warn exc.message
+
+      all.size
     end
 
     private
 
-    # Lazily parse a value, which may be valid JSON or may be a bare string.
-    # TODO: concat a regexp to match JSONable things
-    def process_value(val)
-      JSON.load(val)
-    rescue JSON::ParserError
-      val
-    end
-
-    # Converts the dotted notation to a slashed notation. If a @prefix is set, it applies the prefix.
-    # @example
-    #   "common.proxy.endpoints" => common/proxy/endpoints (or) shard403/common/proxy/endpoints
-    def dot_to_slash(key)
-      key = "#{@prefix}.#{key}" unless @prefix.nil?
-      key.split('.').join('/')
+    # Given a key's relative path, return its absolute REST path in the consul
+    # kv, including any prefix that was specified at startup.
+    def path_to(key)
+      p = '/v1/kv'
+      p << prefix unless prefix.empty?
+      p << key
+      p
     end
   end
 end